+++ JetUncertainty.cc
@@ -44,9 +44,9 @@ void JetUncertainty::Initialize(std::string FileName) {
             // Found a definition line - update current formula
 
             nvar = atoi(Parts[0].c_str());
-            if (Parts.size() <= nvar + 1) continue;
+            if (Parts.size() <= (unsigned long)(nvar + 1)) continue;
             npar = atoi(Parts[nvar + 1].c_str());
-            if (Parts.size() <= nvar + 1 + npar + 1) continue;
+            if (Parts.size() <= (unsigned long)(nvar + 1 + npar + 1)) continue;
 
             CurrentFormula = Parts[nvar + 1 + npar + 1];
             CurrentBinTypes.clear();
@@ -62,7 +62,7 @@ void JetUncertainty::Initialize(std::string FileName) {
         else {
             // Otherwise it's a line with actual JECs, add it to the list
 
-            if (Parts.size() < nvar * 2 + 1) continue;
+            if (Parts.size() < (unsigned long)(nvar * 2 + 1)) continue;
             BinTypes.push_back(CurrentBinTypes);
             std::vector<double> Ranges;
             for (int i = 0; i < nvar * 2; i++) {
